First stable version after merging with V3_2_2

+++ src/OCCViewer/OCCViewer_ViewModel.h
@@ -102,8 +102,8 @@ public:
   bool                            isTrihedronVisible() const;
   virtual void                    setTrihedronShown( const bool );
 
-  int                             trihedronSize() const;
-  virtual void                    setTrihedronSize( const int );
+  double                          trihedronSize() const;
+  virtual void                    setTrihedronSize( const double );
 
 public slots:
   void                            onClearViewAspects();
